pyproject.toml
[tool.poetry]
name = "micro-shop-fa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standart"], version = "^0.35.0"}
pydentic = {extras = ["email"], version = "^0.0.1.dev3"}
pydantic-settings = "^2.10.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.42"}
asyncpg = "^0.30.0"
dotenv = "^0.9.9"
alembic = "^1.16.4"
gunicorn = "^23.0.0"
taskiq-fastapi = "^0.3.5"
taskiq-aio-pika = "^0.4.3"
faststream = "^0.5.48"
aiosmtplib = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
